+++ R/tidy_xlsx.R
@@ -169,5 +169,5 @@ tidy_xlsx <- function(path, sheets = NA) {
     }
   }
   sheets <- standardise_sheet(sheets, all_sheets)
-  xlsx_read_(path, sheets$index, sheets$name, sheets$comments_path)
+  xlsx_read_(path, sheets$sheet_path, sheets$name, sheets$comments_path)
 }