+++ R/utils.R
@@ -39,10 +39,10 @@ excel_format <- function(path) {
 
 standardise_sheet <- function(sheets, all_sheets) {
   if (is.numeric(sheets)) {
-    if (max(sheets) > nrow(all_sheets)) {
-      stop("Only ", nrow(all_sheets), " sheet(s) found.", call. = FALSE)
+    if (max(sheets) > max(all_sheets$index)) {
+      stop("Only ", max(all_sheets$index), " sheet(s) found.", call. = FALSE)
     }
-    all_sheets[sheets, ]
+    all_sheets[all_sheets$index == sheets, ]
   } else if (is.character(sheets)) {
     indices <- match(sheets, all_sheets$name)
     if (anyNA(indices)) {
@@ -59,6 +59,5 @@ standardise_sheet <- function(sheets, all_sheets) {
 xlsx_sheets <- function(path) {
   out <- xlsx_sheets_(path)
   out <- out[order(out$index), ]
-  out$index <- seq_along(out$index)
   out
 }