+++ tests/testthat/test-compatibility.R
@@ -11,8 +11,7 @@ test_that("can import examples-gnumeric.xlsx", {
 })
 
 test_that("can read document from google doc", {
-  # All I want to check is that it doesn't crash, but there's no expect_survival
-  expect_silent(tidy_xlsx("iris-google-doc.xlsx"))
+  expect_error(tidy_xlsx("iris-google-doc.xlsx"), NA)
 })
 
 test_that("gives informative error for a JMP export", {